+++ tools/src/main/java/org/apache/kafka/tools/ThroughputThrottler.java
@@ -13,9 +13,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */
+ **/
 
-package org.apache.kafka.common.utils;
+package org.apache.kafka.tools;
 
 
 /**
